Appveyor: Change 3rd. party OpenSSL Windows binaries used.
Switch to use the ones published at
http://www.npcglib.org/~stathis/blog/precompiled-openssl/

The ones from slproweb.com are built with VS2013 now but we are building
FreeTDS with VS 2008, 2010 and 2015.

So we have the following dependency chain:

- FreeTDS (both as static libs and dynamic libs)
- OpenSSL (as dynamic libs, because that's how we are building FreeTDS)
- The Visual C runtime libs (as dynamic libs, because that's how we are
  building FreeTDS and OpenSSL)

When the C runtime libs are linked in dynamically, a mismatch of
versions is a recipe for disaster. So we need to make sure FreeTDS and
OpenSSL are built using the same MS toolchain version.